% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_analysis}
\alias{cluster_analysis}
\title{Perform Hierarchical Clustering Analysis}
\usage{
cluster_analysis(data, n_clusters = 3, method = "ward.D2")
}
\arguments{
\item{data}{A data frame containing environmental variables (tmax, tmin, RH, etc.)}

\item{n_clusters}{Number of clusters to create (default: 3)}

\item{method}{Clustering method (default: "ward.D2")}
}
\value{
A list containing clustering results and visualizations
}
\description{
This function performs hierarchical clustering on environmental variables
to identify distinct climate scenarios that may influence disease development.
}
\examples{
data(rbsd_data)
cluster_result <- cluster_analysis(rbsd_data[, c("tmax", "tmin", "RH")])
}
