% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_model.R
\name{derive_cluster_rules}
\alias{derive_cluster_rules}
\title{Derive Cluster Rules from Data}
\usage{
derive_cluster_rules(data, cluster_col, feature_cols)
}
\arguments{
\item{data}{Data frame with cluster assignments}

\item{cluster_col}{Name of column containing cluster assignments}

\item{feature_cols}{Names of feature columns}
}
\value{
List of cluster rules
}
\description{
Derive Cluster Rules from Data
}
\keyword{internal}
