% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{plot_cluster_boxplots}
\alias{plot_cluster_boxplots}
\title{Plot Cluster Boxplots}
\usage{
plot_cluster_boxplots(x, data)
}
\arguments{
\item{x}{An object of class "rbsd_cluster" from cluster_analysis()}

\item{data}{Original data frame used for clustering}
}
\value{
Boxplots showing variable distribution by cluster
}
\description{
Plot Cluster Boxplots
}
\examples{
data(rbsd_data)
cluster_result <- cluster_analysis(rbsd_data[, c("tmax", "tmin", "RH")])
plot_cluster_boxplots(cluster_result, rbsd_data[, c("tmax", "tmin", "RH")])
}
