% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{plot_dendrogram}
\alias{plot_dendrogram}
\title{Plot Dendrogram from Clustering Analysis}
\usage{
plot_dendrogram(x, ...)
}
\arguments{
\item{x}{An object of class "rbsd_cluster" from cluster_analysis()}

\item{...}{Additional parameters passed to plot()}
}
\value{
A dendrogram plot
}
\description{
Plot Dendrogram from Clustering Analysis
}
\examples{
data(rbsd_data)
cluster_result <- cluster_analysis(rbsd_data[, c("tmax", "tmin", "RH")])
plot_dendrogram(cluster_result)
}
