% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_model.R
\name{predict_cluster_from_rules}
\alias{predict_cluster_from_rules}
\title{Predict Cluster from Rules}
\usage{
predict_cluster_from_rules(tmax, tmin, RH, rules)
}
\arguments{
\item{tmax}{Maximum temperature}

\item{tmin}{Minimum temperature}

\item{RH}{Relative humidity}

\item{rules}{Cluster rules from derive_cluster_rules()}
}
\value{
Predicted cluster ID
}
\description{
Predict Cluster from Rules
}
\keyword{internal}
