% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict_rbsd}
\alias{predict_rbsd}
\title{Predict RBSD Using Trained Models}
\usage{
predict_rbsd(model, newdata, type = c("A", "PDI"))
}
\arguments{
\item{model}{Trained model from train_ml_models() or train_hybrid_model()}

\item{newdata}{New data for prediction}

\item{type}{Type of prediction ("A" or "PDI")}
}
\value{
Predictions for the specified target
}
\description{
This function makes predictions using either machine learning models
or the hybrid model.
}
\examples{
data(rbsd_data)
processed <- preprocess_data(rbsd_data)
ml_models <- train_ml_models(processed)
predictions <- predict_rbsd(ml_models, processed$test_features, "A")
}
