% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.rbsd_hybrid_model}
\alias{print.rbsd_hybrid_model}
\title{Print Method for rbsd_hybrid_model Objects}
\usage{
\method{print}{rbsd_hybrid_model}(x, ...)
}
\arguments{
\item{x}{An object of class "rbsd_hybrid_model"}

\item{...}{Additional parameters passed to print()}
}
\description{
Print Method for rbsd_hybrid_model Objects
}
