% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.rbsd_processed}
\alias{print.rbsd_processed}
\title{Print Method for rbsd_processed Objects}
\usage{
\method{print}{rbsd_processed}(x, ...)
}
\arguments{
\item{x}{An object of class "rbsd_processed"}

\item{...}{Additional parameters passed to print()}
}
\description{
Print Method for rbsd_processed Objects
}
