% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_risk_map}
\alias{plot_risk_map}
\title{Plot Risk Map for RBSD}
\usage{
plot_risk_map(data, model, type = c("A", "PDI"), resolution = 50)
}
\arguments{
\item{data}{Data frame containing environmental variables}

\item{model}{Trained model for prediction}

\item{type}{Type of prediction ("A" or "PDI")}

\item{resolution}{Resolution of the risk map (default: 50)}
}
\value{
A risk map visualization
}
\description{
This function creates a visualization of RBSD risk based on environmental conditions.
}
\examples{
data(rbsd_data)
processed <- preprocess_data(rbsd_data)
ml_models <- train_ml_models(processed)
plot_risk_map(rbsd_data[, c("tmax", "tmin", "RH")], ml_models, "A")
}
